plots-requirements.txt
certifi==2021.5.30
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.5.4
six==1.16.0
